Code is now organzed. dbListEntry was created for polymorphic reasons.
--

TODO: each activity implements the getData for the ListFragment to present the data. However, in order to present every Hive associated with an Apiary, an argument must be passed (i.e. apiary_id) in the ListFragment.

Suggestion: perhaps create a two unit vector whose values could be null but could also contain the arguments if necessary. As such polymorphy won't be broken.